test: add exec-any-shells test
This test ensure that child_process.exec can work with any shells.
Giving a shell name if $path is defined, or full path is given.

Testing with cmd, powershell, and git-bash, but
the test fails when testing with git-bash. Git-bash is usually placed
at 'C:\Program Files\Git\bin\', and it has a space that cause
the failure.