deps, v8: strip out overly broad DLL Export annotations
This change fixes the ARM64 Windows build by removing the more broad
of two conflicting DLL Export annotations in two places.

The conflict is because both LLVM and MSVC forbid having both a class
and a class member marked for export.

Preferring the more narrowly targeted annotation avoids needing the
LLVM-specific linker flag /Fc:dllexportInlines- to avoid exporting
inline members of classes. When such a function is inlined and
references inline members of other classes, the member of the other
class is referenced and not inlined, causing missing symbol errors
later (since the function is otherwise always inlined, references to it
do not resolve).

A similar problem has been fixed before in V8:

https://chromium-review.googlesource.com/c/v8/v8/+/1478216